+++ src/index.ts
@@ -30,7 +30,7 @@ const main = async () => {
       console.log(pc.red('Specify one (and only one) markdown file.'));
     console.log(pc.yellow('Usage: markdown-gpt-translator [options] <file>'));
     console.log(parser.help());
-    console.log('Docs: https://github.com/smikitky/markdown-gpt-translator\n');
+    console.log('Docs: https://github.com/smikitky/chatgpt-md-translator\n');
     return;
   }