+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This will be the first version released on NPM. It's now published as a standali
 
 **BREAKING**
 
+- Renamed the package/repository name from 'markdown-gpt-translator' to 'chatgpt-md-translator'.
 - This tool no longer reads config data from environment variables (except for `HTTPS_PROXY`). Settings must be written in the config file.
 - Changed a config key name from `GPT_TRANSLATOR_BASE_DIR` to `BASE_DIR`.